+++ packages/expo-video/android/src/main/java/expo/modules/video/VideoView.kt
@@ -114,17 +114,19 @@ open class VideoView(context: Context, appContext: AppContext, useTextureView: B
         VideoManager.onVideoPlayerDetachedFromView(it, this)
       }
       val oldPlayer = videoPlayer
+      val hasEmittedFirstFrame = newPlayer?.firstFrameEventGenerator?.hasSentFirstFrameForCurrentMediaItem
+        ?: false
       oldPlayer?.removeListener(this)
       newPlayer?.addListener(this)
       field = newPlayer
-      shouldHideSurfaceView = !(newPlayer?.hasRenderedAFrameOfVideoSource ?: false)
+      shouldHideSurfaceView = !hasEmittedFirstFrame
       applySurfaceViewVisibility()
       attachPlayer()
       newPlayer?.let {
         VideoManager.onVideoPlayerAttachedToView(it, this)
       }
       if (oldPlayer != newPlayer) {
-        oldPlayer?.hasBeenDisconnectedFromPlayerView()
+        oldPlayer?.hasBeenDisconnectedFromVideoView()
       }
     }
 
@@ -219,7 +221,7 @@ open class VideoView(context: Context, appContext: AppContext, useTextureView: B
   }
 
   fun attachPlayer() {
-    videoPlayer?.changePlayerView(playerView)
+    videoPlayer?.changeVideoView(this)
   }
 
   fun exitFullscreen() {
@@ -306,9 +308,11 @@ open class VideoView(context: Context, appContext: AppContext, useTextureView: B
   }
 
   override fun onRenderedFirstFrame(player: VideoPlayer) {
-    shouldHideSurfaceView = false
-    applySurfaceViewVisibility()
-    onFirstFrameRender(Unit)
+    if (player.currentVideoView == this) {
+      shouldHideSurfaceView = false
+      applySurfaceViewVisibility()
+      onFirstFrameRender(Unit)
+    }
   }
 
   override fun requestLayout() {